updating note name in dialog was added

+++ src/components/Note.js
@@ -1,18 +1,18 @@
-import React, { useState } from "react";
-import cn from "classnames";
-import { useDrag } from "react-dnd";
-import { connect } from "react-redux";
-import Box from "@material-ui/core/Box";
-import { styled, makeStyles } from "@material-ui/core/styles";
-import DeleteNoteBtn from "./buttons/DeleteNoteBtn";
-import EditNoteNameBtn from "./buttons/EditNoteNameBtn";
-import ItemTypesForReactDND from "./ItemTypesForReactDND";
+import React, { useState } from "react"
+import cn from "classnames"
+import { useDrag } from "react-dnd"
+import { connect } from "react-redux"
+import Box from "@material-ui/core/Box"
+import { styled, makeStyles } from "@material-ui/core/styles"
+import DeleteNoteBtn from "./buttons/DeleteNoteBtn"
+import EditNoteNameBtn from "./buttons/EditNoteNameBtn"
+import ItemTypesForReactDND from "./ItemTypesForReactDND"
 
 import {
   getSelectedNoteIdForEditing,
   getIsUserPressEditNoteNameBtn,
-  getFolders
-} from "../redux/store";
+  getFolders,
+} from "../redux/store"
 import {
   changeNoteName,
   saveNoteId,
@@ -21,14 +21,14 @@ import {
   changeNotesInThisFolder,
   changeNoteNameInNotesInThisFolder,
   deleteNoteInNotesInThisFolderOnDragEnd,
-  saveSelectedItemId
-} from "../redux/actions";
+  saveSelectedItemId,
+} from "../redux/actions"
 
 const NoteWrapper = styled(Box)({
   display: "flex",
   width: 300,
-  alignItems: "flex-end"
-});
+  alignItems: "flex-end",
+})
 
 const useStyles = makeStyles(() => ({
   editNoteNameInput: {
@@ -41,10 +41,10 @@ const useStyles = makeStyles(() => ({
       fontSize: 16,
       width: "78vw",
       padding: 5,
-      paddingLeft: 20
+      paddingLeft: 20,
     },
     ["@media (max-width:500px)"]: {// eslint-disable-line no-useless-computed-key
-      width: "90%"
+      width: "90%",
     },
   },
   noteLabel: {
@@ -57,30 +57,30 @@ const useStyles = makeStyles(() => ({
     ["@media (max-width:599px)"]: {// eslint-disable-line no-useless-computed-key
       fontSize: 16,
       padding: "10px 0 10px 20px",
-      minWidth: "73vw"
+      minWidth: "73vw",
     },
     ["@media (max-width:500px)"]: {// eslint-disable-line no-useless-computed-key
-      minWidth: "69vw"
+      minWidth: "69vw",
     },
     ["@media (max-width:440px)"]: {// eslint-disable-line no-useless-computed-key
-      minWidth: "62vw"
+      minWidth: "62vw",
     },
     ["@media (max-width:360px)"]: {// eslint-disable-line no-useless-computed-key
       minWidth: "58vw",
-      paddingLeft: 10
-    }
-  }
-}));
+      paddingLeft: 10,
+    },
+  },
+}))
 
 const style = {
   backgroundColor: "white",
-  width: "fit-content"
-};
+  width: "fit-content",
+}
 
 const styleForMobile = {
   backgroundColor: "white",
-  width: "100%"
-};
+  width: "100%",
+}
 
 const Note = ({
   note,
@@ -95,56 +95,54 @@ const Note = ({
   deleteNoteInNotesInThisFolderOnDragEnd,
   folders,
   saveSelectedItemId,
-  mobile
+  mobile,
 }) => {
-  const classes = useStyles();
-  const { noteName, noteId } = note;
-  const [newNoteName, setNewNoteName] = useState("");
+  const classes = useStyles()
+  const { noteName, noteId } = note
+  const [newNoteName, setNewNoteName] = useState("")
 
-  const handleInputChange = value => {
-    setNewNoteName(value);
-  };
+  const handleInputChange = (value) => {
+    setNewNoteName(value)
+  }
 
-  const item = { ...note, type: ItemTypesForReactDND.BOX };
+  const item = { ...note, type: ItemTypesForReactDND.BOX }
   const [{ opacity }, drag] = useDrag({
     item,
     end(item, monitor) {
-      const dropResult = monitor.getDropResult();
+      const dropResult = monitor.getDropResult()
       if (item && dropResult) {
         const isDropAllowed =
           dropResult.allowedDropEffect === "any" ||
-          dropResult.allowedDropEffect === dropResult.dropEffect;
+          dropResult.allowedDropEffect === dropResult.dropEffect
         if (isDropAllowed) {
           // alert(
           //   `You ${dropResult.dropEffect} noteName - ${noteName} - ${noteId}into ${dropResult.name}!`
           // );
           const folderIdOnDragEnd = monitor.getDropResult().name;
           const droppedFolder = folders.find((folder, i) => {
-            let needFolder = null;
+            let needFolder = null
             if (i > 0) {
               needFolder = folder.notesInThisFolder.some(
-                note => note.noteId === noteId
-              );
+                (note) => note.noteId === noteId
+              )
             }
-            return needFolder;
-          });
+            return needFolder
+          })
 
-          if (droppedFolder) {
-            if (droppedFolder.folderId) {
-              deleteNoteInNotesInThisFolderOnDragEnd(
-                droppedFolder.folderId,
-                noteId
-              );
-            }
+          if (droppedFolder && droppedFolder.folderId) {
+            deleteNoteInNotesInThisFolderOnDragEnd(
+              droppedFolder.folderId,
+              noteId
+            )
           }
-          changeNotesInThisFolder(folderIdOnDragEnd, note);
+          changeNotesInThisFolder(folderIdOnDragEnd, note)
         }
       }
     },
-    collect: monitor => ({
-      opacity: monitor.isDragging() ? 0.4 : 1
-    })
-  });
+    collect: (monitor) => ({
+      opacity: monitor.isDragging() ? 0.4 : 1,
+    }),
+  })
 
   if (selectedNoteIdForEditing === noteId && isUserPressEditNoteNameButton) {
     return (
@@ -156,15 +154,16 @@ const Note = ({
           placeholder="New Note"
           defaultValue={noteName}
           className={classes.editNoteNameInput}
-          onChange={e => handleInputChange(e.target.value)}
+          onChange={(e) => handleInputChange(e.target.value)}
           onBlur={() => {
-            changeNoteName(newNoteName, noteId);
-            changeNoteNameInNotesInThisFolder(newNoteName, noteId);
-            pressEditNoteNameBtn(false);
+            pressEditNoteNameBtn(false)
+            if (newNoteName === "") return
+            changeNoteNameInNotesInThisFolder(newNoteName, noteId)
+            changeNoteName(newNoteName, noteId)
           }}
         />
       </div>
-    );
+    )
   } else {
     return (
       <div
@@ -178,7 +177,7 @@ const Note = ({
             saveSelectedItemId(noteId);
           }}
           className={cn("note-li ", {
-            selected: selectedNoteIdForEditing === noteId
+            selected: selectedNoteIdForEditing === noteId,
           })}
         >
           <NoteWrapper>
@@ -190,28 +189,27 @@ const Note = ({
       </div>
     );
   }
-};
+}
 
-const mapStateToProps = state => ({
+const mapStateToProps = (state) => ({
   isUserPressEditNoteNameButton: getIsUserPressEditNoteNameBtn(state),
   selectedNoteIdForEditing: getSelectedNoteIdForEditing(state),
+  folders: getFolders(state),
+})
 
-  folders: getFolders(state)
-});
-
-const mapDispatchToProps = dispatch => ({
-  pressEditNoteNameBtn: value => dispatch(pressEditNoteNameBtn(value)),
+const mapDispatchToProps = (dispatch) => ({
+  pressEditNoteNameBtn: (value) => dispatch(pressEditNoteNameBtn(value)),
   changeNoteName: (noteName, noteId) =>
     dispatch(changeNoteName(noteName, noteId)),
-  saveNoteId: noteId => dispatch(saveNoteId(noteId)),
-  openDialog: value => dispatch(openDialog(value)),
+  saveNoteId: (noteId) => dispatch(saveNoteId(noteId)),
+  openDialog: (value) => dispatch(openDialog(value)),
   changeNotesInThisFolder: (folderIdOnDragEnd, note) =>
     dispatch(changeNotesInThisFolder(folderIdOnDragEnd, note)),
   changeNoteNameInNotesInThisFolder: (noteName, noteId) =>
     dispatch(changeNoteNameInNotesInThisFolder(noteName, noteId)),
   deleteNoteInNotesInThisFolderOnDragEnd: (folderId, noteId) =>
     dispatch(deleteNoteInNotesInThisFolderOnDragEnd(folderId, noteId)),
-  saveSelectedItemId: itemId => dispatch(saveSelectedItemId(itemId))
-});
+  saveSelectedItemId: (itemId) => dispatch(saveSelectedItemId(itemId)),
+})
 
-export default connect(mapStateToProps, mapDispatchToProps)(Note);
+export default connect(mapStateToProps, mapDispatchToProps)(Note)
